feat: TS - Add URL attribute functionality to load remote views
This change allows TokenScripts to use the url attribute of a view to
specify that the view should be loaded from a remote location such as IPFS.
For bigger applications, it becomes hard for the developer to embed all required
assets inside a small TokenScript; such as fonts, large graphics and other media.

This provides a different route that make it easier to make richer applications
while ensuring TokenScript TSML size remains small. It also demonstrates the
potential for existing DApps to be easily ported to TokenScript, provided
that Web3TokenView.java is updated to provide the full ethereum provider to the client.
Access to the full provider can be specified as a TokenScript view attribute,
adding an extra layer of security.